# Conflicts:
#	README.md
#	azure-pipelines.yml
#	frontend/src/InteractiveImport/Interactive/InteractiveImportModalContent.js
#	frontend/src/Settings/MediaManagement/MediaManagement.js
#	src/NzbDrone.Api/Config/MediaManagementConfigResource.cs
#	src/NzbDrone.Core/Download/Clients/Flood/Flood.cs
#	src/NzbDrone.Core/Download/Clients/Flood/FloodProxy.cs
#	src/NzbDrone.Core/Download/Clients/Flood/FloodSettings.cs
#	src/NzbDrone.Core/Jobs/TaskManager.cs
#	src/NzbDrone.Core/Localization/Core/en.json
#	src/NzbDrone.Core/Notifications/Plex/Server/PlexServerProxy.cs
#	src/NzbDrone.Core/Notifications/Plex/Server/PlexServerService.cs
#	src/NzbDrone.Windows/Disk/DiskProvider.cs
#	src/Radarr.Api.V3/Config/MediaManagementConfigResource.cs
#	src/Radarr.Api.V3/Movies/MovieModule.cs

+++ src/NzbDrone.Common/Disk/DiskProviderBase.cs
@@ -46,7 +46,7 @@ namespace NzbDrone.Common.Disk
         {
             CheckFolderExists(path);
 
-            var dirFiles = GetFiles(path, SearchOption.AllDirectories).ToList();
+            var dirFiles = GetFiles(path, true).ToList();
 
             if (!dirFiles.Any())
             {
@@ -65,7 +65,7 @@ namespace NzbDrone.Common.Disk
 
         private void CheckFolderExists(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             if (!FolderExists(path))
             {
@@ -75,7 +75,7 @@ namespace NzbDrone.Common.Disk
 
         private void CheckFileExists(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             if (!FileExists(path))
             {
@@ -93,19 +93,19 @@ namespace NzbDrone.Common.Disk
 
         public bool FolderExists(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
             return Directory.Exists(path);
         }
 
         public bool FileExists(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
             return FileExists(path, PathStringComparison);
         }
 
         public bool FileExists(string path, StringComparison stringComparison)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             switch (stringComparison)
             {
@@ -125,12 +125,12 @@ namespace NzbDrone.Common.Disk
 
         public bool FolderWritable(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             try
             {
                 var testPath = Path.Combine(path, "radarr_write_test.txt");
-                var testContent = string.Format("This file was created to verify if '{0}' is writable. It should've been automatically deleted. Feel free to delete it.", path);
+                var testContent = $"This file was created to verify if '{path}' is writable. It should've been automatically deleted. Feel free to delete it.";
                 WriteAllText(testPath, testContent);
                 File.Delete(testPath);
                 return true;
@@ -144,35 +144,44 @@ namespace NzbDrone.Common.Disk
 
         public bool FolderEmpty(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             return Directory.EnumerateFileSystemEntries(path).Empty();
         }
 
-        public string[] GetDirectories(string path)
+        public IEnumerable<string> GetDirectories(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
-            return Directory.GetDirectories(path);
+            return Directory.EnumerateDirectories(path, "*", new EnumerationOptions
+            {
+                AttributesToSkip = FileAttributes.System,
+                IgnoreInaccessible = true
+            });
         }
 
-        public string[] GetFiles(string path, SearchOption searchOption)
+        public IEnumerable<string> GetFiles(string path, bool recursive)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
-            return Directory.GetFiles(path, "*.*", searchOption);
+            return Directory.EnumerateFiles(path, "*", new EnumerationOptions
+            {
+                AttributesToSkip = FileAttributes.System,
+                RecurseSubdirectories = recursive,
+                IgnoreInaccessible = true
+            });
         }
 
         public long GetFolderSize(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
-            return GetFiles(path, SearchOption.AllDirectories).Sum(e => new FileInfo(e).Length);
+            return GetFiles(path, true).Sum(e => new FileInfo(e).Length);
         }
 
         public long GetFileSize(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             if (!FileExists(path))
             {
@@ -180,18 +189,37 @@ namespace NzbDrone.Common.Disk
             }
 
             var fi = new FileInfo(path);
+
+            try
+            {
+                // If the file is a symlink, resolve the target path and get the size of the target file.
+                if (fi.Attributes.HasFlag(FileAttributes.ReparsePoint))
+                {
+                    var targetPath = fi.ResolveLinkTarget(true)?.FullName;
+
+                    if (targetPath != null)
+                    {
+                        fi = new FileInfo(targetPath);
+                    }
+                }
+            }
+            catch (IOException ex)
+            {
+                Logger.Trace(ex, "Unable to resolve symlink target for {0}", path);
+            }
+
             return fi.Length;
         }
 
         public void CreateFolder(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
             Directory.CreateDirectory(path);
         }
 
         public void DeleteFile(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
             Logger.Trace("Deleting file: {0}", path);
 
             RemoveReadOnly(path);
@@ -201,8 +229,8 @@ namespace NzbDrone.Common.Disk
 
         public void CloneFile(string source, string destination, bool overwrite = false)
         {
-            Ensure.That(source, () => source).IsValidPath();
-            Ensure.That(destination, () => destination).IsValidPath();
+            Ensure.That(source, () => source).IsValidPath(PathValidationType.CurrentOs);
+            Ensure.That(destination, () => destination).IsValidPath(PathValidationType.CurrentOs);
 
             if (source.PathEquals(destination))
             {
@@ -219,8 +247,8 @@ namespace NzbDrone.Common.Disk
 
         public void CopyFile(string source, string destination, bool overwrite = false)
         {
-            Ensure.That(source, () => source).IsValidPath();
-            Ensure.That(destination, () => destination).IsValidPath();
+            Ensure.That(source, () => source).IsValidPath(PathValidationType.CurrentOs);
+            Ensure.That(destination, () => destination).IsValidPath(PathValidationType.CurrentOs);
 
             if (source.PathEquals(destination))
             {
@@ -237,8 +265,8 @@ namespace NzbDrone.Common.Disk
 
         public void MoveFile(string source, string destination, bool overwrite = false)
         {
-            Ensure.That(source, () => source).IsValidPath();
-            Ensure.That(destination, () => destination).IsValidPath();
+            Ensure.That(source, () => source).IsValidPath(PathValidationType.CurrentOs);
+            Ensure.That(destination, () => destination).IsValidPath(PathValidationType.CurrentOs);
 
             if (source.PathEquals(destination))
             {
@@ -256,14 +284,19 @@ namespace NzbDrone.Common.Disk
 
         public void MoveFolder(string source, string destination, bool overwrite = false)
         {
-            Ensure.That(source, () => source).IsValidPath();
-            Ensure.That(destination, () => destination).IsValidPath();
+            Ensure.That(source, () => source).IsValidPath(PathValidationType.CurrentOs);
+            Ensure.That(destination, () => destination).IsValidPath(PathValidationType.CurrentOs);
 
             Directory.Move(source, destination);
         }
 
         protected virtual void MoveFileInternal(string source, string destination)
         {
+            if (File.Exists(destination))
+            {
+                throw new FileAlreadyExistsException("File already exists", destination);
+            }
+
             File.Move(source, destination);
         }
 
@@ -293,24 +326,25 @@ namespace NzbDrone.Common.Disk
 
         public void DeleteFolder(string path, bool recursive)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
-            var files = Directory.GetFiles(path, "*.*", recursive ? SearchOption.AllDirectories : SearchOption.TopDirectoryOnly);
-            Array.ForEach(files, RemoveReadOnly);
+            var files = GetFiles(path, recursive);
+
+            files.ToList().ForEach(RemoveReadOnly);
 
             Directory.Delete(path, recursive);
         }
 
         public string ReadAllText(string filePath)
         {
-            Ensure.That(filePath, () => filePath).IsValidPath();
+            Ensure.That(filePath, () => filePath).IsValidPath(PathValidationType.CurrentOs);
 
             return File.ReadAllText(filePath);
         }
 
         public void WriteAllText(string filename, string contents)
         {
-            Ensure.That(filename, () => filename).IsValidPath();
+            Ensure.That(filename, () => filename).IsValidPath(PathValidationType.CurrentOs);
             RemoveReadOnly(filename);
 
             // File.WriteAllText is broken on net core when writing to some CIFS mounts
@@ -326,7 +360,7 @@ namespace NzbDrone.Common.Disk
 
         public void FolderSetLastWriteTime(string path, DateTime dateTime)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             if (dateTime.Before(DateTimeExtensions.Epoch))
             {
@@ -338,7 +372,7 @@ namespace NzbDrone.Common.Disk
 
         public void FileSetLastWriteTime(string path, DateTime dateTime)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             if (dateTime.Before(DateTimeExtensions.Epoch))
             {
@@ -363,16 +397,16 @@ namespace NzbDrone.Common.Disk
             }
         }
 
-        public string GetPathRoot(string path)
+        public virtual string GetPathRoot(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             return Path.GetPathRoot(path);
         }
 
         public string GetParentFolder(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             var parent = Directory.GetParent(path.TrimEnd(Path.DirectorySeparatorChar));
 
@@ -419,9 +453,9 @@ namespace NzbDrone.Common.Disk
 
         public void EmptyFolder(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
-            foreach (var file in GetFiles(path, SearchOption.TopDirectoryOnly))
+            foreach (var file in GetFiles(path, false))
             {
                 DeleteFile(file);
             }
@@ -490,12 +524,11 @@ namespace NzbDrone.Common.Disk
 
                 return mounts.Where(drive => drive.RootDirectory.PathEquals(path) ||
                                              drive.RootDirectory.IsParentPath(path))
-                          .OrderByDescending(drive => drive.RootDirectory.Length)
-                          .FirstOrDefault();
+                          .MaxBy(drive => drive.RootDirectory.Length);
             }
             catch (Exception ex)
             {
-                Logger.Debug(ex, string.Format("Failed to get mount for path {0}", path));
+                Logger.Debug(ex, $"Failed to get mount for path {path}");
                 return null;
             }
         }
@@ -509,7 +542,7 @@ namespace NzbDrone.Common.Disk
 
         public List<DirectoryInfo> GetDirectoryInfos(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             var di = new DirectoryInfo(path);
 
@@ -518,18 +551,23 @@ namespace NzbDrone.Common.Disk
 
         public FileInfo GetFileInfo(string path)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             return new FileInfo(path);
         }
 
-        public List<FileInfo> GetFileInfos(string path, SearchOption searchOption = SearchOption.TopDirectoryOnly)
+        public List<FileInfo> GetFileInfos(string path, bool recursive = false)
         {
-            Ensure.That(path, () => path).IsValidPath();
+            Ensure.That(path, () => path).IsValidPath(PathValidationType.CurrentOs);
 
             var di = new DirectoryInfo(path);
 
-            return di.GetFiles("*", searchOption).ToList();
+            return di.EnumerateFiles("*", new EnumerationOptions
+            {
+                AttributesToSkip = FileAttributes.System,
+                RecurseSubdirectories = recursive,
+                IgnoreInaccessible = true
+            }).ToList();
         }
 
         public void RemoveEmptySubfolders(string path)
